Progress #146: Bug fixes
Fix bugs in the infrastructure written so far towards closing this
issue. Also, create a throwaway UI to prove that the decisions are being
loaded correctly.

+++ VotingApp/Apis/DecisionGeneratorApi.hpp
@@ -45,7 +45,7 @@ public:
                                   QObject *parent = 0);
     virtual ~DecisionGeneratorApi() noexcept {}
 
-    Q_INVOKABLE QJSValue getDecisions(uint8_t count);
+    Q_INVOKABLE QJSValue getDecisions(int count);
 
     bool isOutOfDecisions() const { return m_isOutOfDecisions; }
     bool isFetchingDecisions() const { return m_isFetchingDecisions; }
